+++ src/backend/app/agent/streaming/__init__.py
@@ -1,11 +1,11 @@
 """WebSocket payloads and domain → client wire shapes for conversations."""
 
-from app.agent.realtime.conversation_events import (
+from app.agent.streaming.conversation_events import (
     conversation_room,
     emit_conversation_patch,
     emit_to_conversation,
 )
-from app.agent.realtime.wire import conversation_message_to_wire
+from app.agent.streaming.wire import conversation_message_to_wire
 
 __all__ = [
     "conversation_message_to_wire",
