+++ src/backend/app/api/root.py
@@ -9,9 +9,13 @@ from .v1 import play_ground_routes
 from .v1.versioning import router as versioning_router
 # from .v1 import call_routes
 from .v1 import group_routes
-from .v1.agent.workflows import router as agent_workflows_router
-from .v1.conversations import router as conversations_router
-from .v1.conversations import tasks_router as conversation_tasks_router
+from .v1.conversations import chat_router as conversations_chat_router
+from .v1.conversations import conversation_router as conversations_router
+from .v1.conversations import (
+    conversation_tasks_router as conversation_scoped_tasks_router,
+)
+from .v1.conversations import task_router as conversation_tasks_router
+from .v1.conversations import workflow_router as agent_workflows_router
 
 router = APIRouter()
 
@@ -53,5 +57,7 @@ router.include_router(
 router.include_router(group_routes.router, prefix="/groups", tags=["groups"])
 
 router.include_router(conversations_router)
+router.include_router(conversations_chat_router)
+router.include_router(conversation_scoped_tasks_router)
 router.include_router(conversation_tasks_router)
 router.include_router(agent_workflows_router, prefix="/agent")
