+++ src/backend/app/api/v1/conversations/__init__.py
@@ -1,3 +1,15 @@
-from app.api.v1.conversations.routes import router, tasks_router
+from app.api.v1.conversations.routes import (
+    chat_router,
+    conversation_router,
+    conversation_tasks_router,
+    task_router,
+    workflow_router,
+)
 
-__all__ = ["router", "tasks_router"]
+__all__ = [
+    "chat_router",
+    "conversation_router",
+    "conversation_tasks_router",
+    "task_router",
+    "workflow_router",
+]
