+++ src/backend/app/api/v1/conversations/deps.py
@@ -2,19 +2,21 @@
 
 from __future__ import annotations
 
-from fastapi import Request
+from fastapi import Depends, Request
 
-from app.agent.runner.executor import AgentExecutor
+from app.agent.context.graph_traversal import GraphTraversal
 from app.agent.runner.task_manager import TaskManager
 from app.api.dependencies import (
     get_project_conversation_repo as get_conversation_repo,
     get_project_conversation_store as get_conversation_store,
+    get_project_uow,
 )
+from app.db.context import ProjectUoW
 
 __all__ = [
-    "get_agent_executor",
     "get_conversation_repo",
     "get_conversation_store",
+    "get_graph_traversal",
     "get_task_manager",
 ]
 
@@ -26,8 +28,8 @@ def get_task_manager(request: Request) -> TaskManager:
     return tm
 
 
-def get_agent_executor(request: Request) -> AgentExecutor:
-    ex = getattr(request.app.state, "agent_executor", None)
-    if ex is None:
-        raise RuntimeError("agent_executor not initialized on app.state")
-    return ex
+def get_graph_traversal(
+    uow: ProjectUoW = Depends(get_project_uow),
+) -> GraphTraversal:
+    """Graph access scoped to the resolved project, branch, and ref."""
+    return GraphTraversal(uow)
