+++ src/backend/app/api/v1/conversations/routes/__init__.py
@@ -1,0 +1,17 @@
+from app.api.v1.conversations.routes.chat import router as chat_router
+from app.api.v1.conversations.routes.conversation import (
+    router as conversation_router,
+)
+from app.api.v1.conversations.routes.task import (
+    conversation_tasks_router,
+    router as task_router,
+)
+from app.api.v1.conversations.routes.workflows import router as workflow_router
+
+__all__ = [
+    "chat_router",
+    "conversation_router",
+    "conversation_tasks_router",
+    "task_router",
+    "workflow_router",
+]
