+++ src/backend/app/api/v1/conversations/routes/workflows.py
@@ -1,17 +1,23 @@
+"""Background agent workflows (mounted under /api/v1/agent)."""
+
+from __future__ import annotations
+
 from typing import Any, Optional, Type
 
 from fastapi import APIRouter, Depends, HTTPException
 from pydantic import BaseModel
 
 from app.agent.context.graph_traversal import GraphTraversal
-from app.agent.runner.executor import AgentExecutor
+from app.agent.conversation_store import ConversationStore
+from app.agent.service.workflow_service import WorkflowService
 from app.agent.workflows.base import BaseWorkflow
 from app.agent.workflows.description_gen import DescriptionGeneratorWorkflow
 from app.agent.workflows.documentation_gen import DocumentationGeneratorWorkflow
-from app.api.dependencies import get_project_conversation_store
-from app.api.v1.agent.deps import get_agent_executor, get_graph_traversal
-from app.agent.conversation_store import ConversationStore
-
+from app.api.dependencies import (
+    get_project_conversation_store,
+    get_workflow_service,
+)
+from app.api.v1.conversations.deps import get_graph_traversal
 
 router = APIRouter(prefix="/workflows", tags=["Agent Workflows"])
 
@@ -20,8 +26,6 @@ _WORKFLOW_CLASSES: dict[str, Type[BaseWorkflow]] = {
     "description_generator": DescriptionGeneratorWorkflow,
 }
 
-
-# ─── Schemas ──────────────────────────────────────────────
 
 class RunWorkflowRequest(BaseModel):
     workflow_name: str
@@ -36,55 +40,49 @@ class RunWorkflowResponse(BaseModel):
     status: str
 
 
-# ─── Routes ───────────────────────────────────────────────
-
 @router.post("/run", response_model=RunWorkflowResponse, status_code=202)
 async def run_workflow(
     req: RunWorkflowRequest,
-    executor: AgentExecutor = Depends(get_agent_executor),
+    workflow_service: WorkflowService = Depends(get_workflow_service),
     graph: GraphTraversal = Depends(get_graph_traversal),
+    store: ConversationStore = Depends(get_project_conversation_store),
 ):
     """
     Trigger a background workflow (e.g., documentation generation).
     If conversation_id is None, a new conversation is automatically created.
     """
-
     workflow_cls = _WORKFLOW_CLASSES.get(req.workflow_name)
     if not workflow_cls:
         raise HTTPException(
             status_code=400,
-            detail=f"Unknown workflow: {req.workflow_name}"
+            detail=f"Unknown workflow: {req.workflow_name}",
         )
 
     workflow = workflow_cls(
         graph=graph,
-        llm_factory=executor.llm_factory,
+        llm_factory=workflow_service.llm_factory,
     )
 
     try:
         params = dict(req.params or {})
-        # Keep generations separate by route; ignore legacy combined mode params.
         params.pop("mode", None)
 
-        # 2. Instruct executor to start the workflow
-        # The executor handles creating the TaskPart message and submitting to TaskManager
-        conv_id, task_id = await executor.run_workflow(
-            workflow=workflow,
-            conversation_id=req.conversation_id,
+        conv_id, task_id = await workflow_service.run(
+            workflow,
             store=store,
+            conversation_id=req.conversation_id,
             **params,
         )
 
-        # 3. Return accepted status immediately (task is running in background)
         return RunWorkflowResponse(
             conversation_id=conv_id,
             task_id=task_id,
-            status="accepted_and_running"
+            status="accepted_and_running",
         )
 
     except ValueError as e:
-        # E.g., invalid params or conversation_id doesn't exist
-        raise HTTPException(status_code=400, detail=str(e))
+        raise HTTPException(status_code=400, detail=str(e)) from e
     except Exception as e:
         raise HTTPException(
-            status_code=500, detail=f"Failed to start workflow: {e}")
+            status_code=500, detail=f"Failed to start workflow: {e}"
+        ) from e
